refactor: reader 파일 안 구글폼 응답 dto를 별도 클래스 파일로 이동
